doc: consensus in blog mode
Changing the deployment strategy of the consensus series.
This is the first release of the blog.
I've removed the navbar link to the consensus series, but
those who have the link can still see the old content.
Once everything is released as a blog, we can delete the
old content.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>